Remove string highlighting for text blocks and piped texts
Previously, the content of text blocks and piped text lines were being
highlighted as a string, but they aren't strictly strings ('…' or "…")
so the highlight is being removed, is better to consider them just
plain stuff without any type. Also, now they are consistent with the
content of plain HTML tags (i.e. `h1 Text`).

As an extra, pipes are now marked as function names instead of plain
text to make clear that are part of the parent tag, which are also
marked as function names.